modules/orchestration-service/pyproject.toml
[project]
name = "livetranslate-orchestration-service"
version = "1.0.0"
description = "LiveTranslate Orchestration Service - Backend API & Service Coordination"
authors = [
    {name = "LiveTranslate Team", email = "team@livetranslate.com"},
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    # Core web framework
    "fastapi>=0.121.0,<1.0.0",
    "uvicorn[standard]>=0.38.0,<1.0.0",
    # WebSocket support (FastAPI native)
    "websockets>=15.0.1,<16.0.0",
    # HTTP client
    "requests>=2.31.0,<3.0.0",
    "httpx>=0.28.1,<1.0.0",
    # FastAPI dependencies
    "pydantic>=2.12.3,<3.0.0",
    "pydantic-settings>=2.7.1,<3.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    # Configuration
    "pyyaml>=6.0.1,<7.0.0",
    # Data processing
    "python-dateutil>=2.8.2,<3.0.0",
    # Bot management and Google Meet integration
    "google-api-python-client>=2.100.0,<3.0.0",
    "google-auth>=2.23.0,<3.0.0",
    "google-auth-oauthlib>=1.0.0,<2.0.0",
    "google-auth-httplib2>=0.1.0,<1.0.0",
    # Database integration
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.21,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
    # JWT and security
    "pyjwt>=2.8.0,<3.0.0",
    "cryptography>=41.0.4,<42.0.0",
    # HTTP async client
    "aiohttp>=3.8.5,<4.0.0",
    # File watching for config hot-reload
    "watchdog>=3.0.0,<4.0.0",
    # Audio processing (Windows-compatible versions)
    "numpy>=2.2.0,<3.0.0",
    "scipy>=1.16.2,<2.0.0",
    "librosa>=0.11.0,<1.0.0",
    "soundfile>=0.13.1,<1.0.0",
    "pydub>=0.25.1,<1.0.0",
    "ffmpeg-python>=0.2.0,<1.0.0",
    # Enhanced audio processing libraries
    "pedalboard>=0.9.0,<1.0.0",
    "pyloudnorm>=0.1.1,<1.0.0",
    "webrtcvad-wheels>=2.0.11,<3.0.0",
    "pywavelets>=1.4.1,<2.0.0",
    # Redis for session management
    "redis>=6.4.0,<7.0.0",
    # Monitoring and logging
    "prometheus-client>=0.18.0,<1.0.0",
    "psutil>=5.9.5",
    "livetranslate-common",
    # Development dependencies
    "werkzeug>=2.3.7",
    "aiofiles>=24.1.0,<25.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "openvino-genai>=2025.3.0.0,<2026.0.0.0",
    "transformers>=4.57.1,<5.0.0",
    "python-socketio>=5.14.2,<6.0.0",
    "docker>=7.1.0,<8.0.0",
    "pillow>=12.0.0,<13.0.0",
    "opencv-python>=4.12.0.88,<5.0.0.0",
    "greenlet>=3.3.1",
    "reportlab>=4.4.10",
    "openai>=2.26.0",
    "anthropic>=0.84.0",
]

[project.scripts]
start-backend = "src.orchestration_service:main"
start-fastapi = "backend.main:main"
run-tests = "pytest:main"
livedemo = "src.livedemo.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
livetranslate-common = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.0,<1.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "black>=23.9.1",
    "mypy>=1.5.1",
    "pre-commit>=3.4.0",
    "pytest-timeout>=2.4.0",
    "ruff>=0.8.0",
]
test = [
    "httpx>=0.28.1",
    "pytest-mock>=3.15.1",
    "faker>=37.11.0",
    "timecode>=1.4.1",
    "hypothesis>=6.145.1",
    "testcontainers[postgres]>=4.9.0",
    "testcontainers[redis]>=4.9.0",
]
monitoring = [
    "grafana-api>=1.0.3",
    "psutil>=5.9.5",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff config centralized in /ruff.toml at project root

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = "-v --strict-markers"
asyncio_mode = "auto"
pythonpath = ["."]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/conftest.py",
    "*/venv/*",
    "*/virtualenv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
